requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
pydantic-settings==2.3.4
loguru==0.7.2
requests==2.32.3
pathvalidate==3.2.0
cursvalutarbnr==0.0.1
peewee==3.17.6
